Trent Mick: use size_t instead of int where appropriate (in strxfrm(), to hold strlen() outcome).

+++ Modules/_localemodule.c
@@ -281,7 +281,7 @@ PyLocale_strxfrm(self,args)
   PyObject* args;
 {
   char *s,*buf;
-  int n1,n2;
+  size_t n1,n2;
   PyObject *result;
   if(!PyArg_ParseTuple(args,"s:strxfrm",&s))
      return NULL;
